// src/app/page/master/map-diklatjurusan-produk/map-diklatjurusan-produk.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapDiklatjurusanProduk {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}